Added Anim Tree

+++ Assets/Scripts/Animatons/RandomAnimSelector.cs
@@ -12,7 +12,7 @@ namespace Animations
         public override void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
         {
             int randomIndex = Random.Range(m_minAnimIndex, m_maxAnimIndex + 1);
-            animator.SetInteger(m_animVariableName, randomIndex);
+            animator.SetFloat(m_animVariableName, randomIndex);
 
         }
     }
